Chat screen setup

+++ src/services/users.ts
@@ -13,7 +13,10 @@ const getUsers = async (input: string) => {
     }
 };
 
-const deleteFriend = async (friendId: string) => {
+const deleteFriend = async (friendId: string | undefined) => {
+    console.log('DEBUG::deleteFriend', friendId);
+    if (!Boolean(friendId)) throw new Error('No friend to delete.');
+
     const result = await new Client({
         path: `/users/${Session.userId}/friends/${friendId}`
     }).delete();
